pyproject.toml
[project]
name = "activation_keygen"
version = "1.0.0"
dependencies = [
    "ruff == 0.4.1",
    "mypy == 1.10.0",
    "py-machineid == 0.5.1",
    "cryptography == 42.0.5"
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy"
]

[tool.ruff]
line-length = 120
lint.select = [
    "F", "E", "W", "C90", "I", "N", "D", "UP", "YTT", "ANN", "ASYNC", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ",
    "T10", "DJ", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF",
    "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "TD", "FIX", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR",
    "PERF", "RUF",
]
lint.ignore = ["D203", "D213", "FA102", "RET504"]
lint.extend-ignore = [
  "D", "ANN", "FA100", "T201",  # temporario
  "TD003", "FIX002",  # delete
]

[tool.mypy]
strict = true
show_error_codes = true
enable_error_code = "ignore-without-code"
exclude = ["build", ".venv"]